requirements.txt
pytesseract
wand
pillow
pdfminer.six